CI fix
