+++ website/moneybirdsynchronization/management/commands/contactsync.py
@@ -1,4 +1,5 @@
 from django.core.management.base import BaseCommand
+
 from moneybirdsynchronization import services
