+++ website/moneybirdsynchronization/management/commands/statementsync.py
@@ -1,5 +1,3 @@
-import datetime
-
 from django.core.management.base import BaseCommand
 
 from moneybirdsynchronization import services
