+++ website/moneybirdsynchronization/services.py
@@ -1,11 +1,12 @@
 import datetime
 
-from django.template.defaultfilters import date
 from django.db.models import OuterRef, Q, Subquery
+from django.template.defaultfilters import date
 
 from moneybirdsynchronization import emails
-from moneybirdsynchronization.administration import Administration, HttpsAdministration
+from moneybirdsynchronization.administration import Administration
 from moneybirdsynchronization.models import MoneybirdContact
+from moneybirdsynchronization.moneybird import MoneybirdAPIService
 
 from events.models.event import Event
 from members.models import Member
@@ -13,116 +14,40 @@ from payments.models import Payment
 from thaliawebsite import settings
 
 
-class MoneybirdAPIService:
-    def __init__(self):
-        api = HttpsAdministration(
-            settings.MONEYBIRD_API_KEY, settings.MONEYBIRD_ADMINISTRATION_ID
-        )
-        self.api = api
-
-    def update_contact(self, contact):
-        if contact.moneybird_version is None:
-            response = self.api.post("contacts", contact.to_moneybird())
-            contact.moneybird_id = response["id"]
-            contact.moneybird_version = response["version"]
-            contact.save()
-        else:
-            response = self.api.patch(
-                "contacts/{contact.moneybird_id}", contact.to_moneybird()
-            )
-            contact.moneybird_version = response["version"]
-            contact.save()
-
-    def get_financial_account_id(self, identifier):
-        for financial_account in self.api.get("financial_accounts"):
-            if financial_account["identifier"] == identifier:
-                return financial_account["id"]
-        return None
-
-    def get_project_id(self, name):
-        for project in self.api.get("projects"):
-            if project["name"] == name:
-                return project["id"]
-
-        return self.api.post("projects", {"project": {"name": name}})["id"]
-
-    def get_contribution_ledger_id(self):
-        for ledger in self.api.get("ledger_accounts"):
-            if ledger["name"] == "Contribution":
-                return ledger["id"]
-        return None
-
-    def link_transaction_to_financial_account(self, account_id, new_cash_payments):
-        financial_mutations_attributes = []
-        if account_id is None:
-            return
-
-        for instance in new_cash_payments:
-            payment_response = self.api.post(
-                f"external_sales_invoices/{instance.moneybird_invoice_id}/payments",
-                {
-                    "payment": {
-                        "payment_date": instance.created_at.strftime("%Y-%m-%d"),
-                        "price": str(instance.amount),
-                        "financial_account_id": account_id,
-                        "invoice_id": instance.moneybird_invoice_id,
-                    }
-                },
-            )
-
-            financial_mutations_attributes.append(
-                {
-                    "date": instance.created_at.strftime("%Y-%m-%d"),
-                    "amount": str(instance.amount),
-                    "message": instance.topic,
-                }
-            )
-
-        if len(financial_mutations_attributes) > 0:
-            statement_response = self.api.post(
-                "financial_statements",
-                {
-                    "financial_statement": {
-                        "financial_account_id": account_id,
-                        "reference": f"Card payments {datetime.date.today()}",
-                        "financial_mutations_attributes": financial_mutations_attributes,
-                    }
-                },
-            )
-
-            for x in range(len(new_cash_payments)):
-                instance = new_cash_payments[x]
-                instance.moneybird_financial_statement_id = statement_response["id"]
-                instance.moneybird_financial_mutation_id = statement_response[
-                    "financial_mutations"
-                ][x]["id"]
-                instance.save()
-
-                mutation_response = self.api.patch(
-                    f"financial_mutations/{instance.moneybird_financial_mutation_id}/link_booking",
-                    {
-                        "booking_type": "ExternalSalesInvoice",
-                        "booking_id": instance.moneybird_invoice_id,
-                        "price": str(instance.amount),
-                        "description": instance.topic,
-                    },
-                )
+def get_moneybird_api_service():
+    if (
+        settings.MONEYBIRD_ADMINISTRATION_ID is None
+        or settings.MONEYBIRD_API_KEY is None
+    ):
+        return RuntimeError("Moneybird API key or administration ID not set")
+    return MoneybirdAPIService(
+        settings.MONEYBIRD_API_KEY, settings.MONEYBIRD_ADMINISTRATION_ID
+    )
 
 
 def push_thaliapay_batch(instance):
-    apiservice = MoneybirdAPIService()
+    if settings.MONEYBIRD_SYNC_ENABLED is False:
+        return
+
+    apiservice = get_moneybird_api_service()
     payments = Payment.objects.filter(batch=instance)
     tpay_account_id = apiservice.get_financial_account_id("ThaliaPay")
     apiservice.link_transaction_to_financial_account(tpay_account_id, payments)
 
 
 def update_contact(member):
-    apiservice = MoneybirdAPIService()
+    if settings.MONEYBIRD_SYNC_ENABLED is False:
+        return
+
+    apiservice = get_moneybird_api_service()
     apiservice.update_contact(MoneybirdContact.objects.get_or_create(member=member)[0])
 
 
 def delete_contact(instance):
-    apiservice = MoneybirdAPIService()
+    if settings.MONEYBIRD_SYNC_ENABLED is False:
+        return
+
+    apiservice = get_moneybird_api_service()
     member = Member.objects.get(profile=instance)
     contact = MoneybirdContact.objects.get(member=member)
     apiservice.api.delete(f"contacts/{contact.moneybird_id}")
@@ -130,13 +55,18 @@ def delete_contact(instance):
 
 
 def register_event_registration_payment(instance):
-    apiservice = MoneybirdAPIService()
+    if settings.MONEYBIRD_SYNC_ENABLED is False:
+        return
+
+    apiservice = get_moneybird_api_service()
     contact_id = apiservice.api.get(
         f"contacts/customer_id/{settings.MONEYBIRD_UNKOWN_PAYER_ID}"
     )["id"]
     if instance.payment.paid_by is not None:
         try:
-            contact_id = MoneybirdContact.objects.get(member=instance.member).moneybird_id
+            contact_id = MoneybirdContact.objects.get(
+                member=instance.member
+            ).moneybird_id
         except MoneybirdContact.DoesNotExist:
             pass
 
@@ -166,11 +96,16 @@ def register_event_registration_payment(instance):
         response = apiservice.api.post("external_sales_invoices", invoice_info)
         instance.payment.moneybird_invoice_id = response["id"]
         instance.payment.save()
-    except Exception as e:
+    except Administration.Error as e:
         emails.send_sync_error(e, instance.payment)
 
 
 def register_shift_payments(orders, instance):
+    if settings.MONEYBIRD_SYNC_ENABLED is False:
+        return
+
+    apiservice = get_moneybird_api_service()
+
     try:
         event = Event.objects.get(shifts=instance)
         start_date = date(event.start, "Y-m-d")
@@ -179,7 +114,6 @@ def register_shift_payments(orders, instance):
         start_date = date(instance.start, "Y-m-d")
         project_name = f"{instance} [{start_date}]"
 
-    apiservice = MoneybirdAPIService()
     project_id = apiservice.get_project_id(project_name)
 
     for order in orders:
@@ -188,7 +122,9 @@ def register_shift_payments(orders, instance):
         )["id"]
         if order.payer is not None:
             try:
-                contact_id = MoneybirdContact.objects.get(member=order.payer).moneybird_id
+                contact_id = MoneybirdContact.objects.get(
+                    member=order.payer
+                ).moneybird_id
             except MoneybirdContact.DoesNotExist:
                 pass
 
@@ -214,18 +150,24 @@ def register_shift_payments(orders, instance):
             response = apiservice.api.post("external_sales_invoices", invoice_info)
             order.payment.moneybird_invoice_id = response["id"]
             order.payment.save()
-        except Exception as e:
+        except Administration.Error as e:
             emails.send_sync_error(e, instance.payment)
 
 
 def register_food_order_payment(instance):
-    apiservice = MoneybirdAPIService()
+    if settings.MONEYBIRD_SYNC_ENABLED is False:
+        return
+
+    apiservice = get_moneybird_api_service()
+
     contact_id = apiservice.api.get(
         f"contacts/customer_id/{settings.MONEYBIRD_UNKOWN_PAYER_ID}"
     )["id"]
     if instance.payment.paid_by is not None:
         try:
-            contact_id = MoneybirdContact.objects.get(member=instance.member).moneybird_id
+            contact_id = MoneybirdContact.objects.get(
+                member=instance.member
+            ).moneybird_id
         except MoneybirdContact.DoesNotExist:
             pass
 
@@ -255,12 +197,16 @@ def register_food_order_payment(instance):
         response = apiservice.api.post("external_sales_invoices", invoice_info)
         instance.payment.moneybird_invoice_id = response["id"]
         instance.payment.save()
-    except Exception as e:
+    except Administration.Error as e:
         emails.send_sync_error(e, instance.payment)
 
 
 def register_contribution_payment(instance):
-    apiservice = MoneybirdAPIService()
+    if settings.MONEYBIRD_SYNC_ENABLED is False:
+        return
+
+    apiservice = get_moneybird_api_service()
+
     contact_id = apiservice.api.get(
         f"contacts/customer_id/{settings.MONEYBIRD_UNKOWN_PAYER_ID}"
     )["id"]
@@ -294,12 +240,16 @@ def register_contribution_payment(instance):
         response = apiservice.api.post("external_sales_invoices", invoice_info)
         instance.payment.moneybird_invoice_id = response["id"]
         instance.payment.save()
-    except Exception as e:
+    except Administration.Error as e:
         emails.send_sync_error(e, instance.payment)
 
 
 def delete_payment(instance):
-    apiservice = MoneybirdAPIService()
+    if settings.MONEYBIRD_SYNC_ENABLED is False:
+        return
+
+    apiservice = get_moneybird_api_service()
+
     apiservice.api.delete(f"external_sales_invoices/{instance.moneybird_invoice_id}")
     if instance.moneybird_financial_statement_id is not None:
         try:
@@ -323,6 +273,11 @@ def delete_payment(instance):
 
 
 def sync_contacts():
+    if settings.MONEYBIRD_SYNC_ENABLED is False:
+        return
+
+    apiservice = get_moneybird_api_service()
+
     members_without_contact = Member.objects.filter(
         ~Q(
             id__in=Subquery(
@@ -334,8 +289,6 @@ def sync_contacts():
     for member in members_without_contact:
         contact = MoneybirdContact(member=member)
         contact.save()
-
-    apiservice = MoneybirdAPIService()
 
     # fetch contact ids from moneybird
     api_response = apiservice.api.get("contacts")
@@ -373,13 +326,13 @@ def sync_contacts():
         if int(moneybird["id"]) not in ids:
             response = apiservice.api.delete(f"contacts/{moneybird['id']}")
 
+
 def sync_statements():
-    PAYMENT_TYPE_TO_FINANCIAL_ACCOUNT_MAPPING = {
-            Payment.TPAY: "ThaliaPay",
-            Payment.CASH: "cashtanje",
-            Payment.CARD: "pin",
-        }
-    
+    if settings.MONEYBIRD_SYNC_ENABLED is False:
+        return
+
+    apiservice = get_moneybird_api_service()
+
     date = datetime.date.today()
     new_card_payments = Payment.objects.filter(
         type=Payment.CARD,
@@ -396,18 +349,12 @@ def sync_statements():
         moneybird_financial_statement_id=None,
     )
 
-    apiservice = MoneybirdAPIService()
-
     card_account_id = apiservice.get_financial_account_id(
-        PAYMENT_TYPE_TO_FINANCIAL_ACCOUNT_MAPPING[Payment.CARD]
+        settings.PAYMENT_TYPE_TO_FINANCIAL_ACCOUNT_MAPPING["card"]
     )
-    apiservice.link_transaction_to_financial_account(
-        card_account_id, new_card_payments
-    )
+    apiservice.link_transaction_to_financial_account(card_account_id, new_card_payments)
 
     cash_account_id = apiservice.get_financial_account_id(
-        PAYMENT_TYPE_TO_FINANCIAL_ACCOUNT_MAPPING[Payment.CASH]
+        settings.PAYMENT_TYPE_TO_FINANCIAL_ACCOUNT_MAPPING["cash"]
     )
-    apiservice.link_transaction_to_financial_account(
-        cash_account_id, new_cash_payments
-    )
+    apiservice.link_transaction_to_financial_account(cash_account_id, new_cash_payments)
